raw: expose setup and restore to lua, wrapper: use those
It was possible while expecting a char sequence to echo a character due
to the breif amount of time that the terminal is restored. Exposing
setup and restore prevents that.

In addition, pcalling getChar allows us to catch SIGINT and gracefully
handle the error by restoring the terminal before exiting.